refactor: moved banner and item to standalone file
